requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
opencv-contrib-python==4.10.0.84
numpy==2.1.3
Pillow==11.0.0
python-multipart==0.0.12
